third_party/WebKit/LayoutTests/bluetooth/characteristic/service-same-object.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = '[SameObject] test for BluetoothRemoteGATTCharacteristic\'s ' +
  'service.';

bluetooth_test(() => getHealthThermometerDevice()
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('temperature_measurement'))
    .then(characteristic => {
      assert_equals(characteristic.service, characteristic.service);
}), test_desc);
</script>
